+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Version: Added `ShouldBe()` method to assert the version of ScottPlot matches the expected one (#3093)
 * Marker: Added support for `Marker.horizontalBar` to compliment `verticalBar` (#3101) _Thanks @SerhiiMahera_
 * Axis: Span limits are respected when zooming with a window or scroll wheel (#3082) _Thanks @ashe27_
+* Statistics: Added `Descriptive.StdErr()` for calculating standard error of the mean (#3112)
 
 ## ScottPlot 5.0.10-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-12-03_